backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-generativeai>=0.5.0
neo4j>=5.0.0
reportlab>=4.0.0
numpy>=1.24.0
nltk>=3.8.0
